backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
openai==1.12.0
python-dotenv==1.0.0
python-multipart==0.0.6
boto3==1.34.34
Pillow==10.2.0
aiohttp==3.9.3
redis==5.0.1
celery==5.3.6

# Physics Simulation Libraries
pychrono==1.1.0
numpy==1.24.3
scipy==1.10.1
matplotlib==3.7.1

# Video Generation
opencv-python==4.8.0.74
ffmpeg-python==0.2.0
httpx==0.26.0
